# Make Sharing Faster (Again)