--- a/calyx/src/passes/cell_share.rs
+++ b/calyx/src/passes/cell_share.rs
@@ -1,4 +1,7 @@
+use itertools::Itertools;
+
 use super::sharing_components::ShareComponents;
+use crate::analysis::GraphColoring;
 use crate::errors::CalyxResult;
 use crate::{
     analysis::{LiveRangeAnalysis, ReadWriteSet, ShareSet},
@@ -30,6 +33,8 @@ pub struct CellShare {
 
     /// Cell active in continuous assignments, or ref cells (we want to ignore both)
     cont_ref_cells: HashSet<ir::Id>,
+
+    id_to_type: HashMap<ir::Id, ir::CellType>,
 }
 
 impl Named for CellShare {
@@ -50,6 +55,7 @@ impl ConstructVisitor for CellShare {
             live: LiveRangeAnalysis::default(),
             rewrites: HashMap::new(),
             cont_ref_cells: HashSet::new(),
+            id_to_type: HashMap::new(),
             state_shareable,
             shareable,
         })
@@ -59,6 +65,7 @@ impl ConstructVisitor for CellShare {
         self.rewrites = HashMap::new();
         self.live = LiveRangeAnalysis::default();
         self.cont_ref_cells = HashSet::new();
+        self.id_to_type = HashMap::new();
     }
 }
 
@@ -119,5 +126,131 @@ impl ShareComponents for CellShare {
 
     fn get_rewrites(&self) -> &HashMap<ir::Id, ir::RRC<ir::Cell>> {
         &self.rewrites
+    }
+
+    fn set_id_to_type(&mut self, id_to_type: HashMap<ir::Id, ir::CellType>) {
+        self.id_to_type = id_to_type;
+    }
+
+    fn build_conflict_graph(
+        &self,
+        graphs_by_type: &mut HashMap<ir::CellType, GraphColoring<ir::Id>>,
+        c: &ir::Control,
+        is_in_par: bool,
+    ) -> HashMap<&ir::CellType, HashSet<&ir::Id>> {
+        match c {
+            ir::Control::Empty(_) => HashMap::new(),
+            ir::Control::Invoke(ir::Invoke { comp, .. }) => {
+                let mut live_by_type: HashMap<&ir::CellType, HashSet<&ir::Id>> =
+                    HashMap::new();
+                let node_conflicts =
+                    self.lookup_node_conflicts(&comp.clone_name());
+                if !node_conflicts.is_empty() {
+                    for conflict in node_conflicts {
+                        live_by_type
+                            .entry(&self.id_to_type[conflict])
+                            .or_default()
+                            .insert(conflict);
+                    }
+                }
+                for (cell_type, confs) in live_by_type.iter() {
+                    let g = graphs_by_type.get_mut(cell_type).unwrap();
+                    for (a, b) in confs.iter().tuple_combinations() {
+                        g.insert_conflict(a, b)
+                    }
+                }
+                live_by_type
+            }
+            ir::Control::Enable(ir::Enable { group, .. }) => {
+                let mut live_by_type: HashMap<&ir::CellType, HashSet<&ir::Id>> =
+                    HashMap::new();
+                let node_conflicts =
+                    self.lookup_node_conflicts(&group.clone_name());
+                if !node_conflicts.is_empty() {
+                    for conflict in node_conflicts {
+                        live_by_type
+                            .entry(&self.id_to_type[conflict])
+                            .or_default()
+                            .insert(conflict);
+                    }
+                }
+                for (cell_type, confs) in live_by_type.iter() {
+                    let g = graphs_by_type.get_mut(cell_type).unwrap();
+                    for (a, b) in confs.iter().tuple_combinations() {
+                        g.insert_conflict(a, b)
+                    }
+                }
+                live_by_type
+            }
+            ir::Control::Seq(ir::Seq { stmts, .. }) => {
+                let mut acc: HashMap<&ir::CellType, HashSet<&ir::Id>> =
+                    HashMap::new();
+                for stmt in stmts {
+                    let new_confs = self.build_conflict_graph(
+                        graphs_by_type,
+                        stmt,
+                        is_in_par,
+                    );
+                    if is_in_par {
+                        for (cell_type, nodes) in new_confs {
+                            acc.entry(cell_type).or_default().extend(nodes);
+                        }
+                    }
+                }
+                acc
+            }
+            // *Assuming we have removed comb groups*
+            ir::Control::If(ir::If {
+                tbranch, fbranch, ..
+            }) => {
+                let mut tbranch_confs = self.build_conflict_graph(
+                    graphs_by_type,
+                    &*tbranch,
+                    is_in_par,
+                );
+                let fbranch_confs = self.build_conflict_graph(
+                    graphs_by_type,
+                    &*fbranch,
+                    is_in_par,
+                );
+                if is_in_par {
+                    for (cell_type, nodes) in fbranch_confs {
+                        tbranch_confs
+                            .entry(cell_type)
+                            .or_default()
+                            .extend(nodes);
+                    }
+                }
+                tbranch_confs
+            }
+            ir::Control::While(ir::While { body, .. }) => {
+                self.build_conflict_graph(graphs_by_type, &*body, is_in_par)
+            }
+            ir::Control::Par(ir::Par { stmts, .. }) => {
+                let mut acc: HashMap<&ir::CellType, HashSet<&ir::Id>> =
+                    HashMap::new();
+                for stmt in stmts {
+                    let new_confs =
+                        self.build_conflict_graph(graphs_by_type, stmt, true);
+                    for (cell_type, live_cells) in new_confs {
+                        let g = graphs_by_type.get_mut(cell_type).unwrap();
+                        if let Some(conflicting_cells) = acc.get_mut(cell_type)
+                        {
+                            for cell in live_cells.iter() {
+                                for conflict in conflicting_cells.iter() {
+                                    if cell != conflict {
+                                        g.insert_conflict(cell, conflict);
+                                    }
+                                }
+                            }
+                            conflicting_cells.extend(live_cells);
+                        } else {
+                            acc.insert(cell_type, live_cells);
+                        }
+                    }
+                }
+                acc
+            }
+        }
     }
 }