This time I made some more dramatic changes to the structure of Cell Sharing. 

- Since `CellShare` was the only thing that used `ShareComponents` there are is no longer a `CellShare` trait that needs to be implemented .
- I restructured the way I insert the conflicts. Now it is basically performed all in one function that recursively visits each control statement, and inserts the conflicts as it goes through the control statement. 

Another note: I tried to make live ranges be in terms of HashMaps (from cell types to sets of cell names) rather than a HashSet. It wasn't really improving performance all that much, although that may be due to the way I was implementing it. I still have these changes on a separate branch (that I haven't merged), and could play around with it more tomorrow to see if I could get some improvement). 

Improvement Times 
debug
gemm2: .099 s -> .101 s
gemm3: 2.627s -> 2.003 s
gemm4: .905s -> .866 s
gemm6: 104s -> 60 s
gemm8: 28s ->  25 s

release
gemm2: .010 -> .009 s
gemm3: .184s -> .145 s
gemm4: .083s ->  .074 s
gemm6: 5.816s -> 3.025 s
(For some reason I thought that gemm6 was faster when I was testing it in the previous PR but when I tried to run the master branch I got 5.816s). 
gemm8: 1.544s -> 1.303 s